initial program to parse Split Infinity's IU walkthrough. reads lines of text and operates ones a specified terminator is found.
Abstract classes:

ABCParser: base class to handle buffering and action onces a terminator is found
ABCIUParser: just adds some common class variables
ABCSectionParser: ABCParser to handle sections. Split's sections have a """delimiter\nname\ndelimiter""" format. has logic to know if we're in a section description or in the section itself in a method called "_foundterminatorfinish". every class inheriting this one will have to call this method at the end of their implementation of "foundterminator". kinda ugly, but works for my purposes at the moment.

Classes:

SectionParser: parses sections of the walkthrough and looks for the IC section. calls an IC appropriate parser when found.
ICSectionParser: parses users within the IC section and calls item specific parser when found.
ICItemParser: parses items within the user section. for now, just print the parsed items and materials.
IUItems: main class; takes a url (can be either web location or file. uses very simple determination. if string starts with http it's assumed to be a web location) and parses items out of it.